pyproject.toml
[project]
name = "todayos"
version = "0.1.0"
description = "Pseudo-sistema operacional gráfico em Python sobre Linux"
authors = ["Ubuntuinformation <noreply@example.com>"]
dependencies = ["pygame>=2.4.0", "Pillow>=10.0.0"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
